サーバーを起動しなくてもopenapiとapi clientを生成できるようにした
