+++ .gitignore
@@ -7,6 +7,9 @@ packages/web/.next/
 # Hono build output
 packages/api/dist
 
+# Generated OpenAPI schema
+packages/api/openapi.json
+
 # TypeScript
 packages/web/*.tsbuildinfo
