+++ packages/api/package.json
@@ -8,6 +8,7 @@
 		"build": "tsc",
 		"start": "node dist/index.js",
 		"check": "tsc --noEmit",
+		"generate:openapi": "tsx scripts/generate-openapi.ts",
 		"prisma:reset": "prisma migrate reset",
 		"prisma:migrate": "prisma migrate dev",
 		"prisma:studio": "prisma studio",
