+++ packages/api/src/config.ts
@@ -1,0 +1,8 @@
+export const OPENAPI_CONFIG = {
+	openapi: "3.0.0",
+	info: {
+		title: "Meguriai API",
+		version: "1.0.0",
+		description: "Meguriai API Server",
+	},
+} as const;
